updation

+++ src/main/java/com/driver/Main.java
@@ -11,7 +11,7 @@ public class Main {
         Product p = new Product();
         int ans = p.product(10,15);
         int ans1 = p.product(10,15,20);
-        int ans2 = p.product(10.02,10.6);
+        double ans2 = p.product(10.02,10.6);
         System.out.println(ans+" "+ans1+" "+ans2);
     }
